NDArray - changed default type from float64_t to float32_t

+++ src/ndlib/ndarray.h
@@ -29,7 +29,7 @@
 
 namespace laruen::ndlib {
 
-    template <typename T = float64_t>
+    template <typename T = float32_t>
     class NDArray : public ArrayBase {
 
         template <typename> friend class NDArray;
